feat: swagger를 활용해 open API 명세 작성
