+++ src/main/resources/application.properties
@@ -3,3 +3,5 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://database-1.clmeaum2an22.ap-northeast-2.rds.amazonaws.com/toyProject
 spring.datasource.username=admin
 spring.datasource.password=toy12345
+
+spring.mvc.pathmatch.matching-strategy=ant_path_matcher
