+++ src/main/resources/templates/edit.html
@@ -5,8 +5,9 @@
     <title>Edit</title>
 </head>
 <body>
-
-<span id="memberId" th:text="${memberId}"></span><br>
+<h2>회원 정보 수정</h2>
+<span id="memberId" th:text="'수정할 아이디는 ' + ${memberId} + ' 입니다.'"></span><br>
+<p style="text-decoration: underline;">아이디를 동일하게 입력해주세요. (비밀번호가 일치해야 수정됩니다.)</p>
 <!--<form action="/api/user/join" method="post">-->
     <!-- required 속성 적용으로 모든 내용이 채워져야 회원가입 가능   -->
 <form th:action="@{'/api/user/?id=' + ${memberId}}" method="POST">
@@ -28,7 +29,7 @@
     <label for="email">Email: </label>
     <input type="email" id="email" name="memberEmail" required><br><br>
 
-    <input type="submit" value="Submit">
+    <input type="submit" value="수정하기">
 </form>
 <script>
