+++ src/main/resources/templates/index.html
@@ -8,6 +8,5 @@
 <h2>Toy Project</h2>
 <a href="/join">회원 가입</a>
 <a href="/member/search">회원 목록 조회 및 수정</a>
-<a href="/member/edit">회원 정보 수정</a>
 </body>
 </html>
