+++ src/main/resources/templates/join.html
@@ -5,6 +5,7 @@
     <title>Join</title>
 </head>
 <body>
+<h2>회원가입</h2>
 <form action="/api/user/join" method="post">
     <!-- required 속성 적용으로 모든 내용이 채워져야 회원가입 가능   -->
     <label for="id">Id: </label>
@@ -25,7 +26,7 @@
     <label for="email">Email: </label>
     <input type="email" id="email" name="memberEmail" required><br><br>
 
-    <input type="submit" value="Submit">
+    <input type="submit" value="가입하기">
 </form>
 
 </body>
