+++ src/main/resources/templates/search.html
@@ -45,7 +45,7 @@
                     <td>${user.name}</td>
                     <td>${user.phone_number}</td>
                     <td>${user.email}</td>
-                    <td><button onclick="editUser('${user.member_id}')">Edit</button></td>
+                    <td><button onclick="editUser('${user.member_id}')">수정</button></td>
                 `;
             tableBody.appendChild(row);
         });
@@ -57,7 +57,7 @@
         for (let i = 0; i < totalPages; i++) {
             const link = document.createElement('a');
             link.href = '#';
-            link.textContent = i + 1;
+            link.textContent = i + 1 + "페이지        ";
             link.addEventListener('click', () => fetchData(i, 4));
             if (i === currentPage) {
                 link.classList.add('active');
